+++ lib/ember/appkit/rails/sprockets.rb
@@ -1,4 +1,5 @@
 module Ember::Appkit::Rails::Sprockets; end
 
+require 'ember/appkit/rails/sprockets/context'
 require 'ember/appkit/rails/sprockets/processed_asset'
 require 'ember/appkit/rails/sprockets/directive_processor'